'adfs'

+++ QualityControl/FileImportAutomation/ImportQualityControlCases.ps1
@@ -1,10 +1,31 @@
-﻿  Param(
+﻿# Excelproblems - https://superuser.com/questions/579900/why-cant-excel-open-a-file-when-run-from-task-scheduler
+# Excel DCOM https://blogs.technet.microsoft.com/the_microsoft_excel_support_team_blog/2012/11/12/microsoft-excel-or-microsoft-word-does-not-appear-in-dcom-configuration-snap-in/
+Param(
     [Parameter(Mandatory=$true, Position=0)]
-        [string] $newFilePath
-    )
+    [ValidateSet("Production","Development")]
+    [string] $Environment,
+    [Parameter(Mandatory=$false, Position=1)]
+    [string] $DeleteItemsByDataExtractionID,
+    [Parameter(Mandatory=$false, Position=2)]
+    [switch] $DeleteDevelopmentItems
+)
 <##############################################################
 ##                  Helper functions                          #
 ###############################################################>
+function _createBackUpList{
+    Param(
+            [Parameter(Mandatory=$true, Position=0)]
+            [string] $fileBaseName
+        )
+    Import-Module Sharegate	
+
+    $mypassword = ConvertTo-SecureString "MandM7777" -AsPlainText -Force
+    $myusername="nicd@lb.dk"
+    $mysite = Connect-Site -Url $SiteURL -Username $myusername -Password $mypassword
+
+    Copy-List -SourceSite $mysite -Name "Quality Control - Claims Handler Answers" -DestinationSite $mysite -ListTitleUrlSegment $fileBaseName -ListTitle "ShareGate - $fileBaseName"
+}
+
 function _convertStringToDate{
     Param(
     [Parameter(Mandatory=$true, Position=0)]
@@ -49,17 +70,45 @@ param(
      {
         try
         {
-        Remove-PnPListItem -List $listID -Identity $item.Id -Force
+            Remove-PnPListItem -List $listID -Identity $item.Id -Force
         }
         catch
         {
-        Write-Host “error”
+            Write-Host “error”
         }
 
     } #for each end
 
 }
 
+function _deleteItemsByDataExtractionID(){
+param(
+[Parameter(Mandatory=$true)] [string] $listID,
+[Parameter(Mandatory=$true)] [string] $dataExtractionID
+)
+    
+    
+    $items =Get-PnPListItem -List $listID -PageSize 1000
+
+    foreach ($item in $items)
+     {
+        try
+        {
+            #if($item['DataExtractionID'] -eq '01JUL19Q3')
+            if($item['DataExtractionID'] -eq $dataExtractionID)
+            {
+                
+                Remove-PnPListItem -List $listID -Identity $item.Id -Force
+            }
+        }
+        catch
+        {
+            Write-Host “error”
+        }
+
+    } #for each end
+
+}
 
 function _createClaimControlItem(){
 param
@@ -119,14 +168,12 @@ function _readExcelFile{
     (
         [Parameter(Mandatory=$true)] [System.Object] $importFilePath
     )
-
-    $importFilePath = '.\17JUN19_Skadetrans.xlsx'
-
-
+    Write-Output "_readExcelFile.." >> "C:\Git\LBIntranet\QualityControl\FileImportAutomation\log.txt"
     $xlCellTypeLastCell=14
     $startRow=2
     $excel = New-Object -Com Excel.Application
-    $wb = $excel.Workbooks.Open('C:\Git\LBIntranet\QualityControl\FileImportAutomation\17JUN19_Skadetrans.xlsx')
+    $wb = $excel.Workbooks.Open($importFilePath)
+    #$wb = $excel.Workbooks.Open('C:\Git\LBIntranet\QualityControl\FileImportAutomation\17JUN19_Skadetrans.xlsx')
 
 
     $sh = $wb.Sheets.Item(1)
@@ -168,38 +215,72 @@ function _createQuarterlyReport(){
 
 ############################################# START ###################################################
 try{
+    $startClock = Get-Date
+    Write-Output "Starting to import items" >> "C:\Git\LBIntranet\QualityControl\FileImportAutomation\log.txt"
+    $items = Get-ChildItem -Path "C:\Git\LBIntranet\QualityControl\FileImportAutomation" -Filter *.xlsx
+    Write-Output "File read." >> "C:\Git\LBIntranet\QualityControl\FileImportAutomation\log.txt"
 
-$SiteURL = 'https://lbforsikring.sharepoint.com/sites/Skade/'
-
-Connect-PnPOnline -Url $SiteURL -Credentials -NICD-
-
-$QualityControlClaimsHandlerAnswersListID = '433d918b-2e51-4ebb-ab2a-3fc9e2b5c540'
-$DEV_LIST_ID = 'fc98c6c2-1d45-4502-aedd-970f39c474eb'
-#_removeAllListItems -listID $DEV_LIST_ID 
-#return
-$importFilePath = 'C:\Git\LBIntranet\QualityControl\FileImportAutomation\17JUN19_Skadetrans.xlsx'
-$importFilePath = $newFilePath
-
-#$importFilePath = $newFilePath
-Write-Output "This is the path to the new file child script: $newFilePath" >> "C:\Git\LBIntranet\QualityControl\FileImportAutomation\log.txt"
+    if($items.Count -eq 1){
+        
+        
 
 
-_readExcelFile -importFilePath $importFilePath | ForEach-Object{
-    try{
-        Write-Output "Creating ITEM $" >> "C:\Git\LBIntranet\QualityControl\FileImportAutomation\log.txt"
-        $i++
-        Write-Host "counter - " $i
-        _createClaimControlItem -listID $DEV_LIST_ID -itemToCreate $_
+        Write-Output "Fire in the hole." >> "C:\Git\LBIntranet\QualityControl\FileImportAutomation\log.txt"
+        $importFilePath = $items[0].FullName
+        $SiteURL = 'https://lbforsikring.sharepoint.com/sites/Skade/'
+
+        Write-Output "Trying to fly to the Cloud." >> "C:\Git\LBIntranet\QualityControl\FileImportAutomation\log.txt"
+        #Connect-PnPOnline -Url $SiteURL -Credentials $(Get-Credential -Credential:[-NICD-])
+        #Connect-SPOService -Url $o365TenantAdmin -Credential $(Get-StoredCredential -Target [MyLabel])
+        Connect-PnPOnline -Url $SiteURL -Credentials -NICD-
+
+        
+        
+        Write-Output "I'm in the Cloud." >> "C:\Git\LBIntranet\QualityControl\FileImportAutomation\log.txt"
+        <#
+        $QualityControlClaimsHandlerAnswersListID = '433d918b-2e51-4ebb-ab2a-3fc9e2b5c540'
+        $DEV_LIST_ID = 'fc98c6c2-1d45-4502-aedd-970f39c474eb'
+        #>
+        if($Environment -eq "Development"){
+            $QualityControlClaimsHandlerAnswersListID = 'fc98c6c2-1d45-4502-aedd-970f39c474eb'
+            
+            if($DeleteDevelopmentItems.IsPresent -eq $true){
+            _removeAllListItems -listID $QualityControlClaimsHandlerAnswersListID 
+            }
+        }
+        elseif($Environment -eq "Production"){
+            #_createBackUpList -fileBaseName $items.BaseName
+            $QualityControlClaimsHandlerAnswersListID = '433d918b-2e51-4ebb-ab2a-3fc9e2b5c540'
+            if($DeleteItemsByDataExtractionID -ne "")
+            {
+                #_deleteItemsByDataExtractionID -listID $QualityControlClaimsHandlerAnswersListID -dataExtractionID $DeleteItemsByDataExtractionID
+            }
+        }
+        
+
+
+        $date=Get-Date
+        $text = "{0} - Trying to read file {1}" -f $date, $importFilePath
+        Write-Output "Fire in the hole." >> "C:\Git\LBIntranet\QualityControl\FileImportAutomation\log.txt"
+        _readExcelFile -importFilePath $importFilePath | ForEach-Object{
+            try{
+                Write-Output "Creating ITEM $_" >> "C:\Git\LBIntranet\QualityControl\FileImportAutomation\log.txt"
+                $i++
+                _createClaimControlItem -listID $QualityControlClaimsHandlerAnswersListID -itemToCreate $_
+            }
+            catch{
+                Write-Output $_.Exception >> "C:\Git\LBIntranet\QualityControl\FileImportAutomation\log.txt"
+            }
+        }
+
+
+        $stopClock = Get-Date 
+        Write-Host "Kickoff - " $startClock
+        Write-Host "Finished - " $stopClock
     }
-    catch{
-        Write-Output $_.Exception >> "C:\Git\LBIntranet\QualityControl\FileImportAutomation\log.txt"
+    else{
+
     }
-}
-
-
-$stopClock = Get-Date 
-Write-Host "Kickoff - " $startClock
-Write-Host "Finished - " $stopClock
 
 }
 catch{
